refactor/#51: conflict 해결
