--- src/features/post/controllers/post.controller.ts
@@ -1,73 +1,0 @@
-import { Body, Controller, Get, Post, Query } from '@nestjs/common';
-import { CommandBus, QueryBus } from '@nestjs/cqrs';
-import { ApiTags } from '@nestjs/swagger';
-import { CreatePostCommand } from '@features/post/commands/create-post/create-post.command';
-import { ApiPost } from '@features/post/controllers/post.swagger';
-import { CreatePostRequestBodyDto } from '@features/post/dtos/request/create-post.request-body-dto';
-import { FindPostsRequestQueryDto } from '@features/post/dtos/request/find-posts.request-query-dto';
-import { PostResponseDto } from '@features/post/dtos/response/post.response-dto';
-import { PostMapper, PostModel } from '@features/post/mappers/post.mapper';
-import { FindPostsQuery } from '@features/post/queries/find-posts/find-posts.query';
-import { routesV1 } from '@config/app.route';
-import { ApiInternalServerErrorBuilder } from '@libs/api/decorators/api-internal-server-error-builder.decorator';
-import { User } from '@libs/api/decorators/user.decorator';
-import { IdResponseDto } from '@libs/api/dtos/response/id.response-dto';
-import { AggregateID } from '@libs/ddd/entity.base';
-import { SetGuardType } from '@libs/guards/decorators/set-guard-type.decorator';
-import { GuardType } from '@libs/guards/types/guard.constant';
-import { SetPagination } from '@libs/interceptors/pagination/decorators/pagination-interceptor.decorator';
-import { Paginated } from '@libs/types/type';
-
-@ApiTags('Post')
-@ApiInternalServerErrorBuilder()
-@Controller(routesV1.version)
-export class PostController {
-  constructor(
-    private readonly commandBus: CommandBus,
-    private readonly queryBus: QueryBus,
-    private readonly postMapper: PostMapper,
-  ) {}
-
-  @ApiPost.Create({ summary: '게시글 생성 API' })
-  @Post(routesV1.post.root)
-  async create(
-    @User('sub') userId: bigint,
-    @Body() createPostRequestBodyDto: CreatePostRequestBodyDto,
-  ) {
-    const command = new CreatePostCommand({
-      userId,
-      ...createPostRequestBodyDto,
-    });
-
-    const result = await this.commandBus.execute<
-      CreatePostCommand,
-      AggregateID
-    >(command);
-
-    return new IdResponseDto(result);
-  }
-
-  @SetGuardType(GuardType.PUBLIC)
-  @SetPagination()
-  @ApiPost.FindPosts({ summary: '게시글 조회(Pagination) API' })
-  @Get(routesV1.post.root)
-  async findPosts(
-    @Query() findPostsRequestQueryDto: FindPostsRequestQueryDto,
-  ): Promise<[PostResponseDto[], number]> {
-    const query = new FindPostsQuery({
-      ...findPostsRequestQueryDto,
-    });
-
-    const [records, recordCount] = await this.queryBus.execute<
-      FindPostsQuery,
-      Paginated<PostModel>
-    >(query);
-
-    return [
-      records.map((record) =>
-        this.postMapper.toResponseDto(this.postMapper.toEntity(record)),
-      ),
-      recordCount,
-    ];
-  }
-}
