--- src/features/post/controllers/post.swagger.ts
@@ -1,98 +1,0 @@
-import { applyDecorators, HttpStatus } from '@nestjs/common';
-import {
-  ApiBearerAuth,
-  ApiCreatedResponse,
-  ApiOperation,
-} from '@nestjs/swagger';
-import { PostController } from '@features/post/controllers/post.controller';
-import { PostResponseDto } from '@features/post/dtos/response/post.response-dto';
-
-import { IdResponseDto } from '@libs/api/dtos/response/id.response-dto';
-import { HttpBadRequestException } from '@libs/exceptions/client-errors/exceptions/http-bad-request.exception';
-import { HttpUnauthorizedException } from '@libs/exceptions/client-errors/exceptions/http-unauthorized.exception';
-import { COMMON_ERROR_CODE } from '@libs/exceptions/types/errors/common/common-error-code.constant';
-import { OffsetPaginationResponseDto } from '@libs/interceptors/pagination/dtos/offset-pagination-interceptor.response-dto';
-import { CustomValidationError } from '@libs/types/custom-validation-errors.type';
-import { ApiOperator, ApiOperationOptionsWithSummary } from '@libs/types/type';
-
-export const ApiPost: ApiOperator<keyof PostController> = {
-  Create: (
-    apiOperationOptions: ApiOperationOptionsWithSummary,
-  ): MethodDecorator => {
-    return applyDecorators(
-      ApiOperation({
-        ...apiOperationOptions,
-      }),
-      ApiBearerAuth('access-token'),
-      ApiCreatedResponse({
-        description: '게시글 정상적으로 생성됨.',
-        type: IdResponseDto,
-      }),
-      HttpBadRequestException.swaggerBuilder(HttpStatus.BAD_REQUEST, [
-        {
-          code: COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER,
-          description:
-            '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
-          additionalErrors: {
-            errors: [
-              {
-                reason: 'reason',
-                property: 'property',
-                value: 'value',
-              },
-              {
-                reason: 'value',
-                property: 'reere',
-                value: 'valasdue',
-              },
-            ],
-            errorType: CustomValidationError,
-          },
-        },
-      ]),
-      HttpUnauthorizedException.swaggerBuilder(HttpStatus.UNAUTHORIZED, [
-        {
-          code: COMMON_ERROR_CODE.INVALID_TOKEN,
-          description: '유효하지 않은 토큰으로 인해 발생하는 에러',
-        },
-      ]),
-    );
-  },
-
-  FindPosts: (
-    apiOperationOptions: ApiOperationOptionsWithSummary,
-  ): MethodDecorator => {
-    return applyDecorators(
-      ApiOperation({
-        ...apiOperationOptions,
-      }),
-      OffsetPaginationResponseDto.swaggerBuilder(
-        HttpStatus.OK,
-        'posts',
-        PostResponseDto,
-      ),
-      HttpBadRequestException.swaggerBuilder(HttpStatus.BAD_REQUEST, [
-        {
-          code: COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER,
-          description:
-            '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
-          additionalErrors: {
-            errors: [
-              {
-                reason: 'reason',
-                property: 'property',
-                value: 'value',
-              },
-              {
-                reason: 'reason2',
-                property: 'property',
-                value: 'value',
-              },
-            ],
-            errorType: CustomValidationError,
-          },
-        },
-      ]),
-    );
-  },
-};
