--- src/features/post/queries/find-posts/find-posts.query.ts
@@ -1,14 +1,0 @@
-import { PaginatedParams, PaginatedQueryBase } from '@libs/ddd/query.base';
-
-export class FindPostsQuery extends PaginatedQueryBase {
-  readonly title?: string;
-
-  readonly body?: string;
-
-  constructor(props: PaginatedParams<FindPostsQuery>) {
-    super(props);
-
-    this.title = props.title;
-    this.body = props.body;
-  }
-}
