--- src/features/post/repositories/post.repository-port.ts
@@ -1,4 +1,0 @@
-import { PostEntity } from '@features/post/domain/post.entity';
-import { RepositoryPort } from '@libs/ddd/repository.port';
-
-export interface PostRepositoryPort extends RepositoryPort<PostEntity> {}
