--- src/features/post/tokens/di-token.ts
@@ -1,1 +1,0 @@
-export const POST_REPOSITORY_DI_TOKEN = Symbol('POST_REPOSITORY_DI_TOKEN');
