--- src/features/user/domain/user-connection/blog/blog.entity-interface.ts
@@ -1,14 +1,0 @@
-import { AggregateID } from '@libs/ddd/entity.base';
-
-export interface BlogProps {
-  createdBy: AggregateID;
-  connectionId: AggregateID;
-  name: string;
-  deletedAt: Date | null;
-}
-
-export interface CreateBlogProps {
-  createdBy: AggregateID;
-  connectionId: AggregateID;
-  name: string;
-}
