--- src/features/user/domain/user-connection/chat-room/chat-room.entity-interface.ts
@@ -1,14 +1,0 @@
-import { AggregateID } from '@libs/ddd/entity.base';
-
-export interface ChatRoomProps {
-  createdBy: AggregateID;
-  connectionId: AggregateID;
-  name: string;
-  deletedAt: Date | null;
-}
-
-export interface CreateChatRoomProps {
-  createdBy: AggregateID;
-  connectionId: AggregateID;
-  name: string;
-}
