+++ src/features/user/queries/user-connection/find-one-user-connection/find-one-user-connection.query-handler.ts
@@ -1,10 +1,10 @@
 import { TransactionalAdapterPrisma } from '@nestjs-cls/transactional-adapter-prisma';
 import { TransactionHost } from '@nestjs-cls/transactional/dist/src/lib/transaction-host';
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
-import { UserConnectionEntity } from '@features/user/domain/user-connection/user-connection.entity';
-import { UserConnectionMapper } from '@features/user/mappers/user-connection.mapper';
+import { UserConnectionEntity } from '@features/user/user-connection/domain/user-connection.entity';
+import { UserConnectionMapper } from '@features/user/user-connection/mappers/user-connection.mapper';
 import { FindOneUserConnectionQuery } from '@features/user/queries/user-connection/find-one-user-connection/find-one-user-connection.query';
-import { UserConnectionStatus } from '@features/user/types/user.constant';
+import { UserConnectionStatus } from '@features/user/user-connection/types/user.constant';
 import { PrismaService } from '@libs/core/prisma/services/prisma.service';
 import { HttpNotFoundException } from '@libs/exceptions/client-errors/exceptions/http-not-found.exception';
 import { COMMON_ERROR_CODE } from '@libs/exceptions/types/errors/common/common-error-code.constant';
